Move replaceFieldFromNgModelContext() and variablizeContext() static functions to Angular help
